refactor(dashboard): extract remaining shared helpers to providerPageHelpers — #3501 Phase 2b
Moves 22 pure symbols from the 10k-LOC ProviderDetailPageClient god-component
into the acyclic leaf providerPageHelpers.ts, unblocking the extraction of
AddApiKeyModal and EditConnectionModal in a later phase without import cycles.

Symbols moved:
- Web-session credential helpers: getWebSessionCredentialLabel/Hint/CheckLabel,
  getAddCredentialModalTitle
- Upstream-headers helpers + types: upstreamHeadersRecordsEqual, HeaderDraftRow,
  UPSTREAM_HEADERS_UI_MAX, headerRowsToRecord, effectiveUpstreamHeadersForProtocol,
  anyUpstreamHeadersBadge, getProtoSlice; CompatByProtocolMap/CompatModelRow/CompatModelMap
- Codex consts + helpers: CODEX_REASONING_STRENGTH_OPTIONS/ACCOUNT_SERVICE_TIER_VALUES/
  GLOBAL_SERVICE_MODE_VALUES, getCodexServiceTierLabel, normalizeCodexLimitPolicy,
  getCodexRequestDefaults, getClaudeCodeCompatibleRequestDefaults
- Misc: compatProtocolLabelKey, extractCommandCodeCredentialInput,
  normalizeAndValidateHttpBaseUrl, SILICONFLOW_ENDPOINTS, CommandCodeAuthFlowState

New transitive imports wired in the leaf: MODEL_COMPAT_PROTOCOL_KEYS
(@/shared/constants/modelCompat), CodexServiceTier + canonical codex defaults
(@/lib/providers/requestDefaults), CodexGlobalServiceMode (@/lib/providers/codexFastTier),
WebSessionCredentialRequirement (./webSessionCredentials).

Client: 10,288 → 9,980 LOC. Leaf: 589 LOC. All validations green:
typecheck:core 0 errors, smoke test 3/3, check:cycles OK,
helpers unit test 25/25, lint 0 errors.
