+++ src/app/(dashboard)/dashboard/providers/[id]/ProviderDetailPageClient.tsx
@@ -72,11 +72,7 @@ import { maskEmail, pickMaskedDisplayValue, pickDisplayValue } from "@/shared/ut
 import useEmailPrivacyStore from "@/store/emailPrivacyStore";
 import EmailPrivacyToggle from "@/shared/components/EmailPrivacyToggle";
 import ProviderIcon from "@/shared/components/ProviderIcon";
-import {
-  getClaudeCodeCompatibleRequestDefaults as _getClaudeCodeCompatibleRequestDefaults,
-  getCodexRequestDefaults as _getCodexRequestDefaults,
-  type CodexServiceTier,
-} from "@/lib/providers/requestDefaults";
+import { type CodexServiceTier } from "@/lib/providers/requestDefaults";
 import {
   CODEX_FAST_TIER_DEFAULT_SUPPORTED_MODELS,
   getCodexEffectiveServiceTier,
@@ -126,20 +122,35 @@ import {
   providerText,
   providerCountText,
   readBooleanToggle,
+  getWebSessionCredentialLabel,
+  getWebSessionCredentialHint,
+  getWebSessionCredentialCheckLabel,
+  getAddCredentialModalTitle,
+  upstreamHeadersRecordsEqual,
+  UPSTREAM_HEADERS_UI_MAX,
+  headerRowsToRecord,
+  effectiveUpstreamHeadersForProtocol,
+  anyUpstreamHeadersBadge,
+  getProtoSlice,
+  CODEX_REASONING_STRENGTH_OPTIONS,
+  CODEX_ACCOUNT_SERVICE_TIER_VALUES,
+  CODEX_GLOBAL_SERVICE_MODE_VALUES,
+  getCodexServiceTierLabel,
+  normalizeCodexLimitPolicy,
+  getCodexRequestDefaults,
+  getClaudeCodeCompatibleRequestDefaults,
+  compatProtocolLabelKey,
+  extractCommandCodeCredentialInput,
+  normalizeAndValidateHttpBaseUrl,
+  SILICONFLOW_ENDPOINTS,
   type ProviderMessageTranslator,
   type LocalProviderMetadata,
+  type CommandCodeAuthFlowState,
+  type HeaderDraftRow,
+  type CompatByProtocolMap,
+  type CompatModelRow,
+  type CompatModelMap,
 } from "./providerPageHelpers";
-type CompatByProtocolMap = Partial<
-  Record<
-    ModelCompatProtocolKey,
-    {
-      normalizeToolCallId?: boolean;
-      preserveOpenAIDeveloperRole?: boolean;
-      upstreamHeaders?: Record<string, string>;
-    }
-  >
->;
-
 /** PATCH fields for provider model compat (matches API + `ModelCompatPerProtocol` shape). */
 type ModelCompatSavePatch = {
   normalizeToolCallId?: boolean;
@@ -149,33 +160,10 @@ type ModelCompatSavePatch = {
   isHidden?: boolean;
 };
 
-type CompatModelRow = {
-  id?: string;
-  name?: string;
-  source?: string;
-  apiFormat?: string;
-  supportedEndpoints?: string[];
-  normalizeToolCallId?: boolean;
-  preserveOpenAIDeveloperRole?: boolean;
-  isHidden?: boolean;
-  upstreamHeaders?: Record<string, string>;
-  compatByProtocol?: CompatByProtocolMap;
-};
-
-type CompatModelMap = Map<string, CompatModelRow>;
-
 function buildCompatMap(rows: CompatModelRow[]): CompatModelMap {
   const m = new Map<string, CompatModelRow>();
   for (const r of rows) if (r.id) m.set(r.id, r);
   return m;
-}
-
-function getProtoSlice(
-  c: CompatModelRow | undefined,
-  o: CompatModelRow | undefined,
-  protocol: string
-) {
-  return c?.compatByProtocol?.[protocol] ?? o?.compatByProtocol?.[protocol];
 }
 
 function isModelHidden(
@@ -192,90 +180,6 @@ function isModelHidden(
     return Boolean(o.isHidden);
   }
   return false;
-}
-
-function getWebSessionCredentialLabel(
-  t: ProviderMessageTranslator,
-  requirement: WebSessionCredentialRequirement,
-  optional: boolean
-): string {
-  if (requirement.kind === "none") {
-    return providerText(t, "webNoAuthCredentialLabel", "No credential required");
-  }
-  const baseLabel =
-    requirement.kind === "token"
-      ? providerText(t, "webTokenCredentialLabel", "Web session token")
-      : t("sessionCookieLabel");
-  return optional ? `${baseLabel} (${t("optional").toLowerCase()})` : baseLabel;
-}
-
-function getWebSessionCredentialHint(
-  t: ProviderMessageTranslator,
-  requirement: WebSessionCredentialRequirement,
-  providerName: string,
-  editing: boolean
-): string | undefined {
-  if (requirement.kind === "none") return undefined;
-
-  const values = { provider: providerName, credential: requirement.credentialName };
-  if (editing) {
-    return requirement.kind === "token"
-      ? providerText(
-          t,
-          "webTokenEditHint",
-          "Leave blank to keep the current web session token. Credential: {credential}.",
-          values
-        )
-      : providerText(
-          t,
-          "webCookieEditHint",
-          "Leave blank to keep the current session cookie. Required cookie: {credential}.",
-          values
-        );
-  }
-
-  return requirement.kind === "token"
-    ? providerText(
-        t,
-        "webTokenCredentialHint",
-        "Credential: {credential}. Paste the token value from your own signed-in {provider} web session, or a DevTools HAR export if the provider supports it.",
-        values
-      )
-    : providerText(
-        t,
-        "webCookieCredentialHint",
-        "Required cookie: {credential}. Paste the Cookie header value from your own signed-in {provider} web session. Do not include the Cookie: prefix.",
-        values
-      );
-}
-
-function getWebSessionCredentialCheckLabel(
-  t: ProviderMessageTranslator,
-  requirement: WebSessionCredentialRequirement
-): string {
-  if (requirement.kind === "token") return providerText(t, "checkWebToken", "Check token");
-  return providerText(t, "checkCookie", "Check cookie");
-}
-
-function getAddCredentialModalTitle(
-  t: ProviderMessageTranslator,
-  providerName: string,
-  requirement: WebSessionCredentialRequirement | null
-): string {
-  if (!requirement) return t("addProviderApiKeyTitle", { provider: providerName });
-  if (requirement.kind === "none") {
-    return providerText(t, "addProviderConnectionTitle", "Add {provider} connection", {
-      provider: providerName,
-    });
-  }
-  if (requirement.kind === "token") {
-    return providerText(t, "addProviderWebTokenTitle", "Add {provider} web token", {
-      provider: providerName,
-    });
-  }
-  return providerText(t, "addProviderSessionCookieTitle", "Add {provider} session cookie", {
-    provider: providerName,
-  });
 }
 
 function WebSessionCredentialGuide({
@@ -468,34 +372,10 @@ function anyNoPreserveCompatBadge(
   return false;
 }
 
-function upstreamHeadersRecordsEqual(
-  a: Record<string, string>,
-  b: Record<string, string>
-): boolean {
-  const ka = Object.keys(a).sort();
-  const kb = Object.keys(b).sort();
-  if (ka.length !== kb.length) return false;
-  return ka.every((k, i) => k === kb[i] && a[k] === b[k]);
-}
-
-type HeaderDraftRow = { id: string; name: string; value: string };
-
-const UPSTREAM_HEADERS_UI_MAX = 16;
-
 function recordToHeaderRows(rec: Record<string, string>, genId: () => string): HeaderDraftRow[] {
   const entries = Object.entries(rec).filter(([k]) => k.trim());
   if (entries.length === 0) return [{ id: genId(), name: "", value: "" }];
   return entries.map(([name, value]) => ({ id: genId(), name, value }));
-}
-
-function headerRowsToRecord(rows: HeaderDraftRow[]): Record<string, string> {
-  const out: Record<string, string> = {};
-  for (const r of rows) {
-    const k = r.name.trim();
-    if (!k) continue;
-    out[k] = r.value;
-  }
-  return out;
 }
 
 type ProviderModelsApiErrorBody = {
@@ -526,44 +406,6 @@ async function formatProviderModelsErrorResponse(res: Response): Promise<string>
   }
   const st = res.statusText?.trim();
   return st || `HTTP ${res.status}`;
-}
-
-function effectiveUpstreamHeadersForProtocol(
-  modelId: string,
-  protocol: string,
-  customMap: CompatModelMap,
-  overrideMap: CompatModelMap
-): Record<string, string> {
-  const c = customMap.get(modelId);
-  const o = overrideMap.get(modelId);
-  const base: Record<string, string> = {};
-  if (c?.upstreamHeaders && typeof c.upstreamHeaders === "object") {
-    Object.assign(base, c.upstreamHeaders);
-  } else if (o?.upstreamHeaders && typeof o.upstreamHeaders === "object") {
-    Object.assign(base, o.upstreamHeaders);
-  }
-  const pc = getProtoSlice(c, o, protocol);
-  if (pc?.upstreamHeaders && typeof pc.upstreamHeaders === "object") {
-    Object.assign(base, pc.upstreamHeaders);
-  }
-  return base;
-}
-
-function anyUpstreamHeadersBadge(
-  modelId: string,
-  customMap: CompatModelMap,
-  overrideMap: CompatModelMap
-): boolean {
-  const c = customMap.get(modelId);
-  const o = overrideMap.get(modelId);
-  const nonempty = (u: unknown) =>
-    u && typeof u === "object" && !Array.isArray(u) && Object.keys(u as object).length > 0;
-  if (nonempty(c?.upstreamHeaders) || nonempty(o?.upstreamHeaders)) return true;
-  for (const p of MODEL_COMPAT_PROTOCOL_KEYS) {
-    const pc = getProtoSlice(c, o, p);
-    if (nonempty(pc?.upstreamHeaders)) return true;
-  }
-  return false;
 }
 
 interface ModelRowProps {
@@ -829,28 +671,6 @@ interface AddApiKeyModalProps {
   onClose: () => void;
 }
 
-type CommandCodeAuthFlowState = {
-  phase:
-    | "idle"
-    | "starting"
-    | "polling"
-    | "received"
-    | "applying"
-    | "applied"
-    | "expired"
-    | "error";
-  state: string;
-  authUrl: string;
-  callbackUrl: string;
-  expiresAt: string | null;
-  message?: string;
-};
-
-const SILICONFLOW_ENDPOINTS = [
-  { id: "siliconflow", label: "Global", baseUrl: "https://api.siliconflow.com/v1" },
-  { id: "siliconflow-cn", label: "China", baseUrl: "https://api.siliconflow.cn/v1" },
-] as const;
-
 interface EditConnectionModalConnection {
   id?: string;
   name?: string;
@@ -887,74 +707,6 @@ interface EditConnectionModalProps {
   onSave: (data: unknown) => Promise<void | unknown>;
   onClose: () => void;
 }
-const CODEX_REASONING_STRENGTH_OPTIONS = [
-  { value: "none", label: "None" },
-  { value: "low", label: "Low" },
-  { value: "medium", label: "Medium" },
-  { value: "high", label: "High" },
-  { value: "xhigh", label: "XHigh" },
-];
-
-const CODEX_ACCOUNT_SERVICE_TIER_VALUES: CodexServiceTier[] = ["default", "priority", "flex"];
-const CODEX_GLOBAL_SERVICE_MODE_VALUES: CodexGlobalServiceMode[] = [
-  "none",
-  ...CODEX_ACCOUNT_SERVICE_TIER_VALUES,
-];
-
-function getCodexServiceTierLabel(
-  t: ProviderMessageTranslator,
-  value: CodexGlobalServiceMode
-): string {
-  if (value === "none") {
-    return providerText(t, "codexServiceModeNone", "No global setting");
-  }
-  if (value === "default") return providerText(t, "codexServiceTierDefault", "Default");
-  if (value === "priority") return providerText(t, "codexServiceTierPriority", "Priority");
-  return providerText(t, "codexServiceTierFlex", "Flex");
-}
-
-function normalizeCodexLimitPolicy(policy: unknown): { use5h: boolean; useWeekly: boolean } {
-  const record =
-    policy && typeof policy === "object" && !Array.isArray(policy)
-      ? (policy as Record<string, unknown>)
-      : {};
-  return {
-    use5h: typeof record.use5h === "boolean" ? record.use5h : true,
-    useWeekly: typeof record.useWeekly === "boolean" ? record.useWeekly : true,
-  };
-}
-
-/**
- * UI adapter around the canonical getCodexRequestDefaults from requestDefaults.ts.
- * Adds the "medium" fallback for reasoningEffort required by the connection form.
- */
-function getCodexRequestDefaults(providerSpecificData: unknown): {
-  reasoningEffort: string;
-  serviceTier?: CodexServiceTier;
-} {
-  const defaults = _getCodexRequestDefaults(providerSpecificData);
-  return {
-    reasoningEffort: defaults.reasoningEffort ?? "medium",
-    ...(defaults.serviceTier ? { serviceTier: defaults.serviceTier } : {}),
-  };
-}
-
-function getClaudeCodeCompatibleRequestDefaults(providerSpecificData: unknown): {
-  context1m: boolean;
-} {
-  const defaults = _getClaudeCodeCompatibleRequestDefaults(providerSpecificData);
-  return {
-    context1m: defaults.context1m === true,
-  };
-}
-
-function compatProtocolLabelKey(protocol: string): string {
-  if (protocol === "openai") return "compatProtocolOpenAI";
-  if (protocol === "openai-responses") return "compatProtocolOpenAIResponses";
-  if (protocol === "claude") return "compatProtocolClaude";
-  return "compatProtocolOpenAI";
-}
-
 function ModelCompatPopover({
   t,
   effectiveModelNormalize,
@@ -8298,52 +8050,6 @@ function ConnectionRow({
   );
 }
 
-function extractCommandCodeCredentialInput(value: string): string {
-  const trimmed = value.trim();
-  if (!trimmed) return "";
-
-  try {
-    const parsed = JSON.parse(trimmed) as unknown;
-    if (parsed && typeof parsed === "object") {
-      const record = parsed as Record<string, unknown>;
-      const direct = record.apiKey || record.api_key || record.key || record.token;
-      if (typeof direct === "string" && direct.trim()) return direct.trim();
-      const nested = record.data;
-      if (nested && typeof nested === "object") {
-        const nestedRecord = nested as Record<string, unknown>;
-        const nestedKey = nestedRecord.apiKey || nestedRecord.api_key || nestedRecord.key;
-        if (typeof nestedKey === "string" && nestedKey.trim()) return nestedKey.trim();
-      }
-    }
-  } catch {
-    // Not JSON; continue with URL/raw parsing.
-  }
-
-  try {
-    const url = new URL(trimmed);
-    const key =
-      url.searchParams.get("apiKey") ||
-      url.searchParams.get("api_key") ||
-      url.searchParams.get("key") ||
-      url.searchParams.get("token");
-    if (key?.trim()) return key.trim();
-    const hash = url.hash.replace(/^#/, "");
-    if (hash) {
-      const hashParams = new URLSearchParams(hash);
-      const hashKey =
-        hashParams.get("apiKey") ||
-        hashParams.get("api_key") ||
-        hashParams.get("key") ||
-        hashParams.get("token");
-      if (hashKey?.trim()) return hashKey.trim();
-    }
-  } catch {
-    // Not a URL; use the raw value.
-  }
-
-  return trimmed;
-}
-
 function SiliconFlowEndpointModal({
   isOpen,
   onSelect,
@@ -9179,19 +8885,6 @@ function AddApiKeyModal({
       </div>
     </Modal>
   );
-}
-
-function normalizeAndValidateHttpBaseUrl(rawValue, fallbackUrl) {
-  const value = (typeof rawValue === "string" ? rawValue.trim() : "") || fallbackUrl;
-  try {
-    const parsed = new URL(value);
-    if (parsed.protocol !== "http:" && parsed.protocol !== "https:") {
-      return { value: null, error: "Base URL must use http or https" };
-    }
-    return { value, error: null };
-  } catch {
-    return { value: null, error: "Base URL must be a valid URL" };
-  }
 }
 
 function EditConnectionModal({ isOpen, connection, onSave, onClose }: EditConnectionModalProps) {
